+++ packages/create-vexcms/CHANGELOG.md
@@ -1,5 +1,7 @@
 # create-vexcms
 
+## 0.1.0-alpha.8
+
 ## 0.1.0-alpha.7
 
 ### Patch Changes